requirements.txt
# Jamin Industrial Agent runtime dependencies
# Target environments:
# - Windows development workstation
# - Ubuntu production server

# Core data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Configuration and logging
loguru>=0.7.2
PyYAML>=6.0.1
python-dotenv>=1.0.0
jsonschema>=4.21.0
tqdm>=4.66.0
click>=8.1.7
psutil>=5.9.0

# API service
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# Industrial protocol and storage adapters
python-snap7>=1.3.0
pymodbus>=3.5.0
influxdb-client>=1.38.0
apache-iotdb>=1.3.0
aiosqlite>=0.19.0
psycopg[binary]>=3.2.0

# Document and report processing
openpyxl>=3.1.2
PyPDF2>=3.0.1
python-docx>=1.1.0
reportlab>=4.1.0

# Forecasting and machine learning
prophet>=1.1.5
neuralprophet>=0.7.0
statsmodels>=0.14.0
tensorflow-intel>=2.15.0; platform_system == "Windows"
tensorflow>=2.15.0; platform_system == "Linux"

# Knowledge base and vector storage
chromadb>=0.4.22
faiss-cpu>=1.8.0; platform_system == "Linux"

# LLM providers
dashscope>=1.20.0
zhipuai>=2.0.0
openai>=1.14.0

# Ubuntu server only
gunicorn>=21.2.0; platform_system == "Linux"
supervisor>=4.2.5; platform_system == "Linux"

